reorganized frontend dirs a bit.
