+++ frontend/src/edit/patternsTable.js
