+++ frontend/src/map/parcelControl.jsx
@@ -5,11 +5,11 @@ import {
   Button,
 } from '@blueprintjs/core';
 
-import useInterval from './hooks/useInterval';
+import useInterval from '../hooks/useInterval';
 import {
   getLulcTableForParcel,
   postLulcTableForParcel,
-} from './requests';
+} from '../requests';
 
 export default function ParcelControl(props) {
   const {
@@ -17,27 +17,29 @@ export default function ParcelControl(props) {
     addParcel,
     clearSelection,
     sessionID,
+    studyAreaID,
   } = props;
 
-  const [jobID, setJobID] = useState(null);
+  // const [jobID, setJobID] = useState(null);
 
-  useInterval(async () => {
-    const results = await getLulcTableForParcel(jobID);
-    if (!['pending', 'running'].includes(results)) {
-      const addition = {
-        [parcel.parcelID]: {
-          coords: parcel.coords,
-          table: JSON.parse(results).base, // TODO: do we need the scenario identifier in these results?
-        },
-      };
-      setJobID(null);
-      addParcel(addition);
-    }
-  }, (jobID) ? 200 : null); // This server operation should be quick
+  // useInterval(async () => {
+  //   const results = await getLulcTableForParcel(jobID);
+  //   if (!['pending', 'running'].includes(results)) {
+  //     const addition = {
+  //       [parcel.parcelID]: {
+  //         coords: parcel.coords,
+  //         table: JSON.parse(results).base, // TODO: do we need the scenario identifier in these results?
+  //       },
+  //     };
+  //     setJobID(null);
+  //     addParcel(addition);
+  //   }
+  // }, (jobID) ? 200 : null); // This server operation should be quick
 
   const handleClick = async (parcel) => {
-    const jid = await postLulcTableForParcel(sessionID, parcel.coords);
-    setJobID(jid.job_id);
+    // const jid = await postLulcTableForParcel(sessionID, parcel.coords);
+    // setJobID(jid.job_id);
+    postParcel(sessionID, studyAreaID, parcel.coords);
   };
 
   return (
